# docs/openapi.yml
...
openapi: 3.1.0
components:
  schemas:
...
    Instant:
      type: string
      format: date-time
      examples:
      - 2022-03-10T16:15:50Z
...
    Response14:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/TaskDto"
    Response15:
      type: object
      properties:
        task:
          $ref: "#/components/schemas/TaskDto"
    Response16:
      type: object
      properties:
        success:
          type: boolean
...
    TaskDto:
      type: object
      properties:
        id:
          type: string
        featureId:
          type: string
        repositoryId:
          type: string
        title:
          type: string
        slug:
          type: string
        description:
          type: string
        dependsOnTaskId:
          type: string
        implementedAt:
          $ref: "#/components/schemas/Instant"
        createdAt:
          $ref: "#/components/schemas/Instant"
        updatedAt:
          $ref: "#/components/schemas/Instant"
...
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        dependsOnTaskId:
          type: string
        clearDependsOn:
          type: boolean
        implementedAt:
          $ref: "#/components/schemas/Instant"
        clearImplementedAt:
          type: boolean
...
paths:
...
  /projects/api/tasks/{id}:
    put:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTaskRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response14"
        "400":
          description: Bad Request
      summary: Update
      tags:
      - Task Controller
    get:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response15"
      summary: Get
      tags:
      - Task Controller
    delete:
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response16"
      summary: Delete
      tags:
      - Task Controller
info:
  title: qits-projects API
  version: 2026.808.110015